refactor(interpolation) finish configurable interpolation
Pulled interpolators and color maps out into separate modules.
Renamed classes to reflect the fact that they no longer only make dot maps.

BREAKING CHANGE: This changes the name of the classes that render grids into tiles.
  It also changes the way those classes are configured and initialized.
