+++ lib/index.js
@@ -2,14 +2,14 @@
 
 import Choroplethp from './choropleth'
 import ReactChoroplethp from './react/choropleth'
-import Dotp from './dot'
-import ReactDotp from './react/dot'
+import GridTileLayerp from './grid-tile-layer'
+import ReactGridMapLayerp from './react/grid-map-layer'
 import equal from './scales/equal-interval'
 import quantile from './scales/quantile'
 import diverging from './scales/diverging'
 
 export const Choropleth = Choroplethp
 export const ReactChoropleth = ReactChoroplethp
-export const Dot = Dotp
-export const ReactDot = ReactDotp
+export const GridTileLayer = GridTileLayerp
+export const ReactGridMapLayer = ReactGridMapLayerp
 export const scales = { equal, quantile, diverging }
